+++ weaver/processes/builtin/metalink2netcdf.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 """
 Extracts and fetches NetCDF files from a Metalink file containing an URL, and outputs the NetCDF file at a given
 index of the list.