+++ weaver/tweens.py
@@ -53,12 +53,17 @@ def ows_response_tween(request, handler):
         return_error = OWSException(detail=str(err), status=HTTPInternalServerError)
         exc_info_err = sys.exc_info()
         exc_log_lvl = logging.ERROR
+    # FIXME:
+    #   https://github.com/crim-ca/weaver/issues/215
+    #   convivial generation of this repr format should be directly in common exception class
+    raised_err_code = getattr(raised_error, "code", getattr(raised_error, "status_code", 500))
+    raised_err_repr = "({}) <{}> {!s}".format(type(raised_error).__name__, raised_err_code, raised_error)
     if raised_error != return_error:
-        err_msg = "\n  Raised: [{!r}]\n  Return: [{!r}]".format(raised_error, return_error)
+        err_msg = "\n  Raised: [{}]\n  Return: [{!r}]".format(raised_err_repr, return_error)
     else:
-        err_msg = " [{!r}]".format(raised_error)
+        err_msg = " [{}]".format(raised_err_repr)
     LOGGER.log(exc_log_lvl, "Handled request exception:%s", err_msg, exc_info=exc_info_err)
-    LOGGER.debug("Handled request details:\n%s\n%s", str(raised_error), getattr(raised_error, "text", ""))
+    LOGGER.debug("Handled request details:\n%s\n%s", raised_err_repr, getattr(raised_error, "text", ""))
     return return_error